+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Note: the BigBuy API is not versioned and some endpoints are added or removed without notice. The removal of methods
 that call deleted endpoints is not considered a breaking change.
 
-## 3.17.0 (unreleased)
+## 3.17.0 (2022/11/08)
 
 * Document `create_multi_shipping_order` response
 * Rename `BBWarehouseError` as `BBWarehouseSplitError` but keep the alias `BBWarehouseError` for now